apps/api/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-filter>=24.2
drf-spectacular>=0.27
psycopg2-binary>=2.9
python-dotenv>=1.0
pytest>=8.2
pytest-django>=4.8
